% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-automl.R
\name{automl_fit_impl}
\alias{automl_fit_impl}
\title{H2O AutoML Modeling Function (Bridge)}
\usage{
automl_fit_impl(formula, data, ...)
}
\arguments{
\item{formula}{formula specified for training}

\item{data}{A dataframe containing the training data}

\item{...}{Additional arguments.}
}
\description{
H2O AutoML Modeling Function (Bridge)
}
