% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-automl.R
\name{automl_fit_impl}
\alias{automl_fit_impl}
\title{H2O AutoML Modeling Function (Bridge)}
\usage{
automl_fit_impl(x, y, ...)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{...}{Additional arguments passed to \code{h2o.automl()}.}
}
\value{
A fitted model with class \code{automl_fit_impl} and \code{modeltime_bridge}.
}
\description{
H2O AutoML Modeling Function (Bridge)
}
