% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-leaderboard.R
\name{automl_leaderboard}
\alias{automl_leaderboard}
\title{Get the H2O AutoML Leaderboard}
\usage{
automl_leaderboard(object)
}
\arguments{
\item{object}{An object created by \code{\link[=automl_reg]{automl_reg()}}}
}
\value{
A tibble containing the H2O AutoML Leaderboard
}
\description{
The H2O AutoML Leaderboard lists any models that have been created during the \code{automl_reg()}
training process. The training process automatically uses the top model. However, users
can change the model used using \code{automl_change_model()}.
}
\details{
Find out which models have been created. Used to get the Model ID.
}
